+++ plugins/json.ts
@@ -4,25 +4,30 @@ import { merge } from "../core/utils.ts";
 import type { Site } from "../core.ts";
 
 export interface Options {
-  /** The list of extensions used to load data */
-  dataExtensions: string[];
-
-  /** The list of extensions used to load pages */
-  pagesExtensions: string[];
+  /** The list of extensions used to load files */
+  extensions: string[] | {
+    pages: string[];
+    data: string[];
+  };
 }
 
 // Default options
 const defaults: Options = {
-  dataExtensions: [".json"],
-  pagesExtensions: [".tmpl.json"],
+  extensions: {
+    data: [".json"],
+    pages: [".tmpl.json"],
+  },
 };
 
 /** A plugin to add support for JSON files */
 export default function (userOptions?: Partial<Options>) {
   const options = merge(defaults, userOptions);
+  const extensions = Array.isArray(options.extensions)
+    ? { pages: options.extensions, data: options.extensions }
+    : options.extensions;
 
   return (site: Site) => {
-    site.loadData(options.dataExtensions, json);
-    site.loadPages(options.pagesExtensions, json);
+    site.loadData(extensions.data, json);
+    site.loadPages(extensions.pages, json);
   };
 }