+++ plugins/jsx.ts
@@ -6,7 +6,10 @@ import type { Data, Engine, Helper, Site } from "../core.ts";
 
 export interface Options {
   /** The list of extensions this plugin applies to */
-  extensions: string[];
+  extensions: string[] | {
+    pages: string[];
+    components: string[];
+  };
 }
 
 // Default options
@@ -59,11 +62,14 @@ export class JsxEngine implements Engine {
 /** Register the plugin to support JSX and TSX files */
 export default function (userOptions?: Partial<Options>) {
   const options = merge(defaults, userOptions);
+  const extensions = Array.isArray(options.extensions)
+    ? { pages: options.extensions, components: options.extensions }
+    : options.extensions;
 
   return (site: Site) => {
     const engine = new JsxEngine();
 
-    site.loadPages(options.extensions, loader, engine);
-    site.loadComponents(options.extensions, loader, engine);
+    site.loadPages(extensions.pages, loader, engine);
+    site.loadComponents(extensions.components, loader, engine);
   };
 }