+++ plugins/yaml.ts
@@ -5,7 +5,10 @@ import type { Site } from "../core.ts";
 
 export interface Options {
   /** The list of extensions this plugin applies to */
-  extensions: string[];
+  extensions: string[] | {
+    pages: string[];
+    data: string[];
+  };
 }
 
 // Default options
@@ -16,9 +19,12 @@ const defaults: Options = {
 /** A plugin to add support for YAML files */
 export default function (userOptions?: Partial<Options>) {
   const options = merge(defaults, userOptions);
+  const extensions = Array.isArray(options.extensions)
+    ? { pages: options.extensions, data: options.extensions }
+    : options.extensions;
 
   return (site: Site) => {
-    site.loadData(options.extensions, yaml);
-    site.loadPages(options.extensions, yaml);
+    site.loadData(extensions.data, yaml);
+    site.loadPages(extensions.pages, yaml);
   };
 }